+++ tests/open_alchemy/test_integration.py
@@ -412,7 +412,7 @@ def test_database_not_autoincrement(engine, sessionmaker):
 @pytest.mark.integration
 def test_database_many_to_one_relationship(engine, sessionmaker):
     """
-    GIVEN specification with a schema with an object relationship
+    GIVEN specification with a schema with a many to one object relationship
     WHEN schema is created, values inserted in both tables and queried
     THEN the data is returned as it was inserted.
     """
@@ -474,8 +474,8 @@ def test_database_many_to_one_relationship(engine, sessionmaker):
 @pytest.mark.integration
 def test_database_many_to_one_relationship_fk(engine, sessionmaker):
     """
-    GIVEN specification with a schema with an object relationship with a defined
-        foreign key
+    GIVEN specification with a schema with a many to one object relationship with a
+        defined foreign key
     WHEN schema is created, values inserted in both tables and queried
     THEN the data is returned as it was inserted.
     """
@@ -529,6 +529,60 @@ def test_database_many_to_one_relationship_fk(engine, sessionmaker):
     queried_model = session.query(model).first()
     assert queried_model.ref_table_name == "ref table name 1"
     assert queried_model.ref_table.name == "ref table name 1"
+
+
+@pytest.mark.integration
+def test_database_one_to_one_relationship(engine, sessionmaker):
+    """
+    GIVEN specification with a schema with an one to one object relationship
+    WHEN schema is created, values inserted in both tables and queried
+    THEN the data is returned as it was inserted.
+    """
+    # Defining specification
+    spec = {
+        "components": {
+            "schemas": {
+                "RefTable": {
+                    "properties": {
+                        "id": {"type": "integer", "x-primary-key": True},
+                        "name": {"type": "string"},
+                    },
+                    "x-tablename": "ref_table",
+                    "x-backref": "table",
+                    "x-uselist": False,
+                    "type": "object",
+                },
+                "Table": {
+                    "properties": {
+                        "id": {"type": "integer", "x-primary-key": True},
+                        "name": {"type": "string"},
+                        "ref_table": {"$ref": "#/components/schemas/RefTable"},
+                    },
+                    "x-tablename": "table",
+                    "type": "object",
+                },
+            }
+        }
+    }
+    # Creating model factory
+    base = declarative.declarative_base()
+    model_factory = open_alchemy.init_model_factory(spec=spec, base=base)
+    model = model_factory(name="Table")
+    ref_model = model_factory(name="RefTable")
+
+    # Creating models
+    base.metadata.create_all(engine)
+    # Creating instance of model and ref_model
+    ref_model_instance = ref_model(id=11, name="ref table name 1")
+    model_instance = model(id=12, name="table name 1", ref_table=ref_model_instance)
+    session = sessionmaker()
+    session.add(ref_model_instance)
+    session.add(model_instance)
+    session.flush()
+
+    # Querying session
+    queried_ref_model = session.query(ref_model).first()
+    assert queried_ref_model.table.id == 12
 
 
 @pytest.mark.parametrize(