+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## Version _unreleased_ - _no date yet_
+## Version 0.9.0 - 2019-11-10
 
 - Add _from_dict_ and _to_dict_ functions to all models that are used to construct a model from a dictionary and to convert a model instance to a dictionary, respectively.
 - Add _x-foreign-key-column_ extension property to define a custom foreign key constraint for many to one relationships.